added attempt to resolve shortUser at login with autosuggest
if the user logged in with anything but their short user for the first time, the DB would be queried for the shortUser, which obviously didn't exist. this would cause an attempt to login with null. Now, an attempt is made to resolve whatever the user has entered into a shortUser with the autosuggest. partially deals with #68 case 1 with full email