Fix bug: no semesters when Authenticate
semesters would be correctly parsed when toUser was called as part of autoSuggest, but not when toUser was called as part of an Authenticate. The main difference is with autosuggest the bind is made by the resource account and with authenticate the bind is made with the user's credentials. In both cases, the results returned were identical for the memberOf elements. However, the code would try querrying LDAP for the DN of each memberOf element; the courses are in a folder to which users do not have read access. This was the cause of the failure. Fortunately, the DN was already present in the original query reseponse, so the CN and OU of each record could be parsed successfully.