Suppress test case where Adobe makes a faulty print
With adobe, v4 drivers, color, and the custom grayscale toggle, adobe makes a grayscale image which does not have a monochrome declaration but which is not equally coloured. This means that if someone manually sets colour to true and grayscale to true, it will count as color.

The result of the psInfo call is correct: the page is in color. However, I don't want to encode this faulty behaviour as the expected one