add dbquery for students which already have a semester
allows us to not requery those users.

Why not get the ones which need it? because there are actually 3 types of users we can pull up (much like a 3-state buffer): Users which need the semester, Users which already have it, and users which were not in the DB (which need it). So I think it's better to get the list of all eligible, remove those which have already been granted, and then process the rest through a standard merge (authenticationManager will autofetch if absent)