use much more efficient getAllCurrentlyEligible
This takes advantage of the fact that a FullUser retrieved from LDAP has the current semester added if the FullUser is eligible. So if we change the mergeUser to include this granted semester, we can greatly simplify the process. 

We also introduce the benefit of merging by shortUser instead of equality, which is the actually correct thing to do.