fix tests to compensate for .copy not copying _id
So fun story, Kotlin data classes don't copy parameters from classes they inherit from. We could do something with an interface, but then we have to override in all subclasses. We could then use a delegate, but then we end up with some weird stuff happening in the internals of the delegate instantiation causing the JPA interop to fail.